frontend/course: extract assignment from assignment list
